init
-created repository and first project within it
-used node.js "npm" commands to set up a package.json and install packages used in the first project
-wrote the HTML and Javascript for the first project
    -created a "Drone" class
    -created a constructor, standard methods, static methods, and getters and setters for the private "id" property
    -instantiated objects of the "Drone" class and defined their property values
    -did various console.log tests with the two objects of the "Drone" class

You can run the project in a dev environment in the browser by opening a Windows command line, navigating to the project location, and running "npm run dev". See Node.js documentation for more info. Alternatively, refer to the "Advanced Javascript Course" in Pluralsight for detailed instructions on how to get a dev environment set up and running